Rename and reroute guest import review page
